Crate wintls

This is supported on Windows and (x86-64 or x86) only.

Creates static thread-locals for Windows binaries.

Essentially, this is a manual implementation of the thread_local feature.

Note this crate is a playground for my education and not intended for production use. If anything does end up being useful then I’ll create a new crate with just the good parts, or else incorporate it into existing code.

Use

This requires a nightly compiler. You will also need to add #![feature(asm)] to your main.rs or lib.rs file if you use the TLS macros.

Example

#![feature(asm)]

wintls::static_thread_local!{
    static TEST: u32 = 0xfeedface;
}

fn main() {
    println!("{:x}", TEST.get()); // feedface

    TEST.set(5);
    println!("{:x}", TEST.get()); // 5

    std::thread::spawn(|| {
        println!("{:x}", TEST.get()); // feedface
    }).join();
}

Modules

dtor

Register thread local destructors

rawraw

Low level primitives for building static thread-local abstractions.

Macros

get_static

Convenience macro for getting the static thread-local value by its identifier.

init_static

Initialize a static as a thread-local.

set_static

Convenience macro for setting the static thread-local value by its identifier.

static_key

Returns a key that identifies the thread local.

static_ptr

Convenience macro for getting a static thread-local pointer by its identifier.

static_thread_local

Statically initialize a thread local.

unsafe_local

Create an UnsafeLocal.

Structs

StaticThreadLocal

Enables setting or getting a static thread local value.

UnsafeLocal

Grants unsafe access to the thread local.